Refactor : CHANGE FINDWIN.CLASS, INPUT.CLASS, PLAY.CLASS IN CONTROLLER
