+++ src/main/java/racingcar/Application.java
@@ -1,15 +1,6 @@
 package racingcar;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import static racingcar.FindWin.findWinCars;
-import static racingcar.Input.inputCarName;
-import static racingcar.Input.inputNum;
-import static racingcar.Play.findEachCarRandomNum;
-import static racingcar.Play.start;
-import static racingcar.Print.printEachRotate;
-import static racingcar.Print.printResult;
+import static racingcar.Controller.Play.start;
 
 public class Application {
